+++ src/lib/fragments/Cytoscape.react.js
@@ -1,0 +1,336 @@
+/**
+ * JavaScript Requirements: cytoscape
+ * React.js requirements: react-cytoscapejs
+ */
+import React, { Component } from 'react';
+import CytoscapeComponent from 'react-cytoscapejs';
+import _ from 'lodash';
+
+import { propTypes, defaultProps } from '../components/Cytoscape.react';
+
+/**
+A Component Library for Dash aimed at facilitating network visualization in
+Python, wrapped around [Cytoscape.js](http://js.cytoscape.org/).
+ */
+export default class Cytoscape extends Component {
+    constructor(props) {
+        super(props);
+
+        this.handleCy = this.handleCy.bind(this);
+        this._handleCyCalled = false;
+    }
+
+    generateNode(event) {
+        const ele = event.target;
+
+        const isParent = ele.isParent(),
+            isChildless = ele.isChildless(),
+            isChild = ele.isChild(),
+            isOrphan = ele.isOrphan(),
+            renderedPosition = ele.renderedPosition(),
+            relativePosition = ele.relativePosition(),
+            parent = ele.parent(),
+            style = ele.style();
+
+        // Trim down the element objects to only the data contained
+        const edgesData = ele.connectedEdges().map(ele => {
+            return ele.data()
+        }),
+            childrenData = ele.children().map(ele => {
+                return ele.data()
+            }),
+            ancestorsData = ele.ancestors().map(ele => {
+                return ele.data()
+            }),
+            descendantsData = ele.descendants().map(ele => {
+                return ele.data()
+            }),
+            siblingsData = ele.siblings().map(ele => {
+                return ele.data()
+            });
+
+        const { timeStamp } = event;
+        const {
+            classes,
+            data,
+            grabbable,
+            group,
+            locked,
+            position,
+            selected,
+            selectable
+        } = ele.json();
+
+        let parentData;
+        if (parent) {
+            parentData = parent.data();
+        } else {
+            parentData = null;
+        }
+
+        const nodeObject = {
+            // Nodes attributes
+            edgesData,
+            renderedPosition,
+            timeStamp,
+            // From ele.json()
+            classes,
+            data,
+            grabbable,
+            group,
+            locked,
+            position,
+            selectable,
+            selected,
+            // Compound Nodes additional attributes
+            ancestorsData,
+            childrenData,
+            descendantsData,
+            parentData,
+            siblingsData,
+            isParent,
+            isChildless,
+            isChild,
+            isOrphan,
+            relativePosition,
+            // Styling
+            style
+        };
+        return nodeObject;
+    }
+
+
+    generateEdge(event) {
+        const ele = event.target;
+
+        const midpoint = ele.midpoint(),
+            isLoop = ele.isLoop(),
+            isSimple = ele.isSimple(),
+            sourceData = ele.source().data(),
+            sourceEndpoint = ele.sourceEndpoint(),
+            style = ele.style(),
+            targetData = ele.target().data(),
+            targetEndpoint = ele.targetEndpoint();
+
+        const { timeStamp } = event;
+        const {
+            classes,
+            data,
+            grabbable,
+            group,
+            locked,
+            selectable,
+            selected,
+        } = ele.json();
+
+        const edgeObject = {
+            // Edges attributes
+            isLoop,
+            isSimple,
+            midpoint,
+            sourceData,
+            sourceEndpoint,
+            targetData,
+            targetEndpoint,
+            timeStamp,
+            // From ele.json()
+            classes,
+            data,
+            grabbable,
+            group,
+            locked,
+            selectable,
+            selected,
+            // Styling
+            style
+        };
+
+        return edgeObject;
+    }
+
+    handleCy(cy) {
+        // If the cy pointer has not been modified, and handleCy has already
+        // been called before, than we don't run this function.
+        if (cy === this._cy && this._handleCyCalled) {
+            return;
+        }
+        this._cy = cy;
+        window.cy = cy;
+        this._handleCyCalled = true;
+
+        // ///////////////////////////////////// CONSTANTS /////////////////////////////////////////
+        const SELECT_THRESHOLD = 100;
+
+        const selectedNodes = cy.collection();
+        const selectedEdges = cy.collection();
+
+        // ///////////////////////////////////// FUNCTIONS /////////////////////////////////////////
+        const refreshLayout = _.debounce(() => {
+            /**
+             * Refresh Layout if needed
+             */
+            const {
+                autoRefreshLayout,
+                layout
+            } = this.props;
+
+            if (autoRefreshLayout) {
+                cy.layout(layout).run()
+            }
+        }, SELECT_THRESHOLD);
+
+        const sendSelectedNodesData = _.debounce(() => {
+            /**
+             This function is repetitively called every time a node is selected
+             or unselected, but keeps being debounced if it is called again
+             within 100 ms (given by SELECT_THRESHOLD). Effectively, it only
+             runs when all the nodes have been correctly selected/unselected and
+             added/removed from the selectedNodes collection, and then updates
+             the selectedNodeData prop.
+             */
+            const nodeData = selectedNodes.map(el => el.data());
+
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    selectedNodeData: nodeData
+                })
+            }
+        }, SELECT_THRESHOLD);
+
+        const sendSelectedEdgesData = _.debounce(() => {
+            const edgeData = selectedEdges.map(el => el.data());
+
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    selectedEdgeData: edgeData
+                })
+            }
+        }, SELECT_THRESHOLD);
+
+        // /////////////////////////////////////// EVENTS //////////////////////////////////////////
+        cy.on('tap', 'node', event => {
+            const nodeObject = this.generateNode(event);
+
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    tapNode: nodeObject,
+                    tapNodeData: nodeObject.data
+                });
+            }
+        });
+
+        cy.on('tap', 'edge', event => {
+            const edgeObject = this.generateEdge(event);
+
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    tapEdge: edgeObject,
+                    tapEdgeData: edgeObject.data
+                });
+            }
+        });
+
+        cy.on('mouseover', 'node', event => {
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    mouseoverNodeData: event.target.data()
+                })
+            }
+        });
+
+        cy.on('mouseover', 'edge', event => {
+            if (typeof this.props.setProps === 'function') {
+                this.props.setProps({
+                    mouseoverEdgeData: event.target.data()
+                })
+            }
+        });
+
+        cy.on('select', 'node', event => {
+            const ele = event.target;
+
+            selectedNodes.merge(ele);
+            sendSelectedNodesData();
+        });
+
+        cy.on('unselect remove', 'node', event => {
+            const ele = event.target;
+
+            selectedNodes.unmerge(ele);
+            sendSelectedNodesData();
+        });
+
+        cy.on('select', 'edge', event => {
+            const ele = event.target;
+
+            selectedEdges.merge(ele);
+            sendSelectedEdgesData();
+        });
+
+        cy.on('unselect remove', 'edge', event => {
+            const ele = event.target;
+
+            selectedEdges.unmerge(ele);
+            sendSelectedEdgesData();
+        });
+
+        cy.on('add remove', () => {
+            refreshLayout();
+        });
+    }
+
+    render() {
+        const {
+            // HTML attribute props
+            id,
+            style,
+            className,
+            // Common props
+            elements,
+            stylesheet,
+            layout,
+            // Viewport Manipulation
+            pan,
+            zoom,
+            // Viewport Mutability and gesture Toggling
+            panningEnabled,
+            userPanningEnabled,
+            minZoom,
+            maxZoom,
+            zoomingEnabled,
+            userZoomingEnabled,
+            boxSelectionEnabled,
+            autoungrabify,
+            autolock,
+            autounselectify
+        } = this.props;
+
+        return (
+            <CytoscapeComponent
+                id={id}
+                cy={this.handleCy}
+                className={className}
+                style={style}
+                elements={CytoscapeComponent.normalizeElements(elements)}
+                stylesheet={stylesheet}
+                layout={layout}
+                pan={pan}
+                zoom={zoom}
+                panningEnabled={panningEnabled}
+                userPanningEnabled={userPanningEnabled}
+                minZoom={minZoom}
+                maxZoom={maxZoom}
+                zoomingEnabled={zoomingEnabled}
+                userZoomingEnabled={userZoomingEnabled}
+                boxSelectionEnabled={boxSelectionEnabled}
+                autoungrabify={autoungrabify}
+                autolock={autolock}
+                autounselectify={autounselectify}
+            />
+        )
+    }
+}
+
+
+Cytoscape.defaultProps = defaultProps;
+Cytoscape.propTypes = propTypes;